lost-exact: recover 12 provably-lost EXACT functions, triage the rest
Worked the 114-entry lost-EXACT list from max_divergence --history. Recovered 12,
each traced to the commit that dropped it:

  Blowfish_encipher  60.41 -> 100  an added <string.h> and NOTHING else
  SetupSprite        64.21 -> 100  one void* result split into two locals
  ScreenToCell       86.93 -> 100  retail materializes the rect address once
  ClickToggle        94.78 -> 100  an m_tab hoist added with the enum names
  CActionArea::Tick  94.93 -> 100  i64*/i32* pointer locals removed
  Broadcast{One,}Channel{,Table}   a u8 temp collapsed into a direct store
  ParseSerial        99.98 -> 100  RestoreChildren(...,0) retyped to LOGIC_NONE
                                   (= -1) where retail pushes 0
  Load/Resolve/Reload/LoadImageA   the enum campaign dropped static_cast<u32>,
                                   lowering a SIGNED switch ladder

Triaged the remainder honestly: 5 are doctrine (the /Ob1 inline-budget wall, and
one peak that was only ever reachable via a __declspec(naked) hand-asm body that
no longer exists), 23 are regalloc-only artifacts with a source lever probed and
refuted on 12 of them, 6 remain open.

3395 -> 3407 exact, 89.13 -> 89.15% fuzzy. Full gate tier passes with
GRUNTZ_LABELS_ACK unset. stale_markers 9 -> 0. jcc_sieve SIGNEDNESS 4 -> 3.

+++ src/DDrawMgr/DDrawSurfacePair.cpp
@@ -175,7 +175,7 @@ void CDDrawSurfacePair::Unload() {
 RVA(0x00163e50, 0x8b)
 i32 CDDrawSurfacePair::LoadImage(CParseSource* src) {
     FileImageFormat type;
-    switch (src->GetEntryTag()) {
+    switch (static_cast<u32>(src->GetEntryTag())) {
         case IMGTAG_PMB:
             type = FMT_BMP;
             break;
